[AIRFLOW-5200] Move GCP PubSub to core
This commit moves GCP PubSub from contrib to core.
For more information check AIP-21.
